Import achievements as rake task
The current application allows the administrator to add achievements via the
user interface. This is fine, but it involves a lot of clicking around, page
transitions, and copy&paste.

This patch effects the import of achievements sourced from a structured text
file into the DB. The benefits here are that:

* the achievements can be versioned.
* importing achievements now takes mere seconds.
* the achievements can be shared and imported among leagues.
* adding new achievements is as easy as editing as text file.

The following code snippet imports the achievements listed in the YAML file
into the DB.

```bash
$ rake desperado:achievements_import[data/achievements/season_01.yaml]
```

Importing the next season's achievements is as easy as specifying another
achievements file.

```bash
$ rake desperado:achievements_import[data/achievements/bristol_02.yaml]
```

## There Are No Tests

For this first cut I thought I'd just drop it on you Johno and get your
feedback. I'm not a Ruby+Rails engineer so there's probably better/easier ways
of accomplishing this and I'm not keen to polish a turd.

## Not Very Smart

The current implementation merely slurps the YAML and saves the achievement
directly to the DB. There's no smarts here: for example, there is no check
to see if an achievment with the same name already exists.

If more smarts were required I'd cut the achievement importing code out into a
dedicated `AchievementImporter` class/function to enable easier testing and to
keep the rake task small.

I didn't go there immediately because this works and I'd rather get your
feedback earlier before assembling a mountain of turd.