Display side & points on acheev's title bar
There is lots of whitespace on the title bar; pulling this information up onto the bar fills that whitespace and leaves the description area free to display just the description of the acheev.

Again, this change also collapses some of the vertical space and results in a "tighter" grid. (I grok that's subjective though.)

I should think that someone with some UI design chops could further refine the title banner into something cool and polished.